requirements.txt
apache-airflow-providers-apache-spark==4.1.1
hdfs==2.7.3
pyspark
kafka-python